There is a mistake within the mex_files with respect the transfer of data from Matlab to Mex. I found out that when Mex reads in a Matlab matrix (with mxGetPr) it reads the matrix column-wise. While the c-code is written in such a way that it expects row-wise reading.
- For the localization coordinates (A and B), this is solved by correctly indexing the coordinates within the c-code. 
- for the uncertainties (scale_A and scale_B), this is not solved and the sigmas are used wrongly. By plotting the sigma-matrices used for calculating the cost, I found out that the Nx2 matrix of uncertainties is read column-wise assing sigmas to the wrong coordinates. I solved this by changing the indexing 
- Also the rotation matrix is read wrongly, here I also changed the indexing to correct for this. 

Now the mex_functions work and output the correct cost value. However, only for the GPU case (mex_expdist), I could not resolve the issue, since the indexing is deeply rooted in the memory allocation and tiled over multiple threats within the cuda code. This is solved by reformatting the sigma matrix for particle A. This can be done by the function: "correct_uncer" which needs to be used when calling the mex_expdist function. I changed this is pairFitting3D and pairFitting3D_symmetric, which are the only two functions that directly use mex_expdist. For the CPU function (mex_expdist_cpu), I could completely solve the issue by changing the indexing within the cuda files.